CSS-fixes and password generator added.

+++ frontend/src/pages/Messages.js
@@ -11,10 +11,11 @@ import Logout from "../pages/LogOut";
 import DropMenu from "./Dropdown";
 import updateIcon from "../icons/update.png"
 import TimedLogout from "./TimedLogout";
-import alert from "../icons/alert.png";
+import alertIcon from "../icons/alert.png";
 import UsersPresent from "./usersPresent";
 import censor from "../icons/censor.png";
 import MessageCards from "./MessageCards"
+
 
 const Messages = (props) => {
 
@@ -24,29 +25,53 @@ const Messages = (props) => {
     const [lbltext, setLblText] = useState('Disable hover effect')
     const [hoverStatus, setHoverStatus] = useState(true)
     const [hideidAndDate, setHideIdAndDate] = useState(false)
-    const [timedLogOff,setTimedLogOff] = useState(false)
-    const [alignText,setAlingText]=useState("message")
-    const [censored,setSencored]=useState(false)
-    const [showInCards,setShowInCards]=useState(false)
-    const [hideMsgDiv,setHideMsgDiv]=useState(false)
-    const [stars,setStars]=useState(false)
-    
-    const navigate = useNavigate()
-   
+    const [timedLogOff, setTimedLogOff] = useState(false)
+    const [alignText, setAlingText] = useState("message")
+    const [censored, setSencored] = useState(false)
+    const [showInCards, setShowInCards] = useState(false)
+    const [hideMsgDiv, setHideMsgDiv] = useState(false)
+    const [stars, setStars] = useState(false)
+    const [confirm, setConfirm] = useState(false)
+    var first="";
 
- 
-    
+
+    const navigate = useNavigate()
+    /*
+    let date = new Date().toLocaleDateString("fi-FI");
+    var repDate=date.replace(".","/").replace(".","/")
+    var finalDate = repDate
+    finalDate = new Date(finalDate.split('/')[2], finalDate.split('/')[1] - 1, finalDate.split('/')[0]);
+    var date2 = "4/7/2024"
+    date2 = new Date(date2.split('/')[2], date2.split('/')[1] - 1, date2.split('/')[0]);
+    var timeDiff = Math.abs(date2.getTime() - finalDate.getTime());
+    var diffDays = Math.ceil(timeDiff / (1000 * 3600 * 24));
+    /*
+    //let date = new Date().toLocaleDateString("fi-FI");
+    var today = new Date("5/7/2024")
+    var dateobject = new Date("4/7/2024")
+    let Difference_In_Time = today.getTime() - dateobject.getTime();
+    let Difference_In_Days =
+        Math.round
+            (Difference_In_Time / (1000 * 3600 * 24));*/
+
+
+   // console.log(diffDays)
+
+
+
+
+
     //timevar parametri saadan timedlogout komponentista
-    const timedLogout = (timevar) =>{
-        console.log("selected "+timevar)
+    const timedLogout = (timevar) => {
+        console.log("selected " + timevar)
         setTimeout(() => {
             navigate("/")
             //aika annetaan millisekunteina, joten sekuntiarvo täytyy kertoa tuhannella
-          }, timevar*1000);
-           
-            
+        }, timevar * 1000);
+
+
     }
-     //tämä funktio vaihtaa divin tyylimäärittelyä.
+    //tämä funktio vaihtaa divin tyylimäärittelyä.
     const changeHoverStatus = () => {
 
         console.log(hoverStatus)
@@ -102,38 +127,54 @@ const Messages = (props) => {
         }
     }
 
-    const handleDelete = async (id) => {
-        try {
-            await axios.delete("http://localhost:8800/messages/" + id)
-            window.location.reload();
 
-        } catch (err) {
-            console.log(err)
+
+    const handleDelete = async (id, txt) => {
+        //vahvistusikkuna
+        var text = "This will delete a message with ID " + id + "and text " + txt + " are you sure?"
+        window.confirm(text)
+        if (window.confirm() == true) {
+            try {
+                await axios.delete("http://localhost:8800/messages/" + id)
+                window.location.reload();
+
+            } catch (err) {
+
+                console.log(err)
+
+
+            }
 
         }
+
+
+
+
+
+
+
 
     }
 
     //funktio saa parametrina klikatun viestin id:n. p-tagi, jossa teksti näytetään on nimetty m+id
     //yhdistelmällä. näin saadaan ainoastaan valitun viestin tekstiosio piiloon vaihtamassa luokka
     //m+id:stä censored luokkaan.
-    const DoCensor = (id)=>{
+    const DoCensor = (id) => {
         setSencored(!censored)
-        if (censored)
-            {
-                
-                document.getElementById("m"+id).setAttribute("class","censored")
-                document.getElementById("c"+id).innerHTML="CENSORED"
-               
+        if (censored) {
 
-            }
+            document.getElementById("m" + id).setAttribute("class", "censored")
+            document.getElementById("c" + id).innerHTML = "CENSORED"
+
+
+        }
         else {
-            document.getElementById("m"+id).setAttribute("class","notCensored")
-            document.getElementById("c"+id).innerHTML=""
+            document.getElementById("m" + id).setAttribute("class", "notCensored")
+            document.getElementById("c" + id).innerHTML = ""
         }
     }
     //tämä funktio vaihtaa viestien esitystapaa divin ja BS5 korttien välillä.
-    const changeViews=()=>{
+    const changeViews = () => {
         setHideMsgDiv(!hideMsgDiv)
         setShowInCards(!showInCards)
 
@@ -143,59 +184,64 @@ const Messages = (props) => {
     //star-rating funktio, joka muuttaa tähden värin mustasta punaiseen
     //clicks muuttujana avulla siirrytyään star1->eteenpäin ja id:n avulla
     //päivitetään aina yhden viestiosion arvostelua.
-    const rate = (id)=>{
-       
-        clicks+=1
-       
-        document.getElementById("star"+clicks+id).setAttribute("class","fa fa-star checked")
-        if (clicks==5)
-            {
-                clicks=0
-            }
-       
-       
+    const rate = (id) => {
+
+        clicks += 1
+
+        document.getElementById("star" + clicks + id).setAttribute("class", "fa fa-star checked")
+
+        if (clicks == 5) {
+            clicks = 0
         }
 
-        //tähtien resetointi
-    const resetStars=(id)=>{
-        console.log(id)
-        
-        document.getElementById("star1"+id).setAttribute("class","fa fa-star")
-        document.getElementById("star2"+id).setAttribute("class","fa fa-star")
-        document.getElementById("star3"+id).setAttribute("class","fa fa-star")
-        document.getElementById("star4"+id).setAttribute("class","fa fa-star")
-        document.getElementById("star5"+id).setAttribute("class","fa fa-star")
+
     }
-     
-        const starsSelected=()=>{
-            setStars(!stars)
-            
-          
-        }
-    
-      
-       
-    
-   
 
+    //tähtien resetointi
+    const resetStars = (id) => {
+        console.log(id)
+
+        document.getElementById("star1" + id).setAttribute("class", "fa fa-star")
+        document.getElementById("star2" + id).setAttribute("class", "fa fa-star")
+        document.getElementById("star3" + id).setAttribute("class", "fa fa-star")
+        document.getElementById("star4" + id).setAttribute("class", "fa fa-star")
+        document.getElementById("star5" + id).setAttribute("class", "fa fa-star")
+    }
+
+    const starsSelected = () => {
+        setStars(!stars)
+
+
+    }
+
+    const endOfPage = () => {
+        const scrollingElement = (document.scrollingElement || document.body);
+        scrollingElement.scrollTop = scrollingElement.scrollHeight;
+
+    }
     return (
 
         <div className="firstMsgDiv" >
-             
-            <UsersPresent/>
+
+            <UsersPresent />
             <Logout />
-            
+
+
             {/*välitetään timedlogout komponentille timedlogout funktio, eli käytetään
             messages komponentissa olevaa funktiota toisesta funktiosta*/}
-            <TimedLogout timedLogout={timedLogout}/>
+            <TimedLogout timedLogout={timedLogout} />
             {/*message komponentin statemuuttujan välitys dropmenu komponentille*/}
-            <DropMenu setHideIdAndDate={setHideIdAndDate} hideidAndDate={hideidAndDate} setHoverOff={setHoverOff} hoverOff={hoverOff} />
-            <label htmlFor="hoverOff">{lbltext}</label>
-            <input id="hoverOff" type="checkbox" onClick={changeHoverStatus} onChange={() => setHoverStatus(!hoverStatus)}></input><br></br>
-            <label htmlFor="bscards">Show messages in Bootstrap cards</label>
-            <input id="bscards" type="checkbox" onChange={changeViews}></input>
+            <DropMenu setHideIdAndDate={setHideIdAndDate} hideidAndDate={hideidAndDate} setHoverOff={setHoverOff} hoverOff={hoverOff} endOfPage={endOfPage} />
+            <span className="changeCB">
+                <label htmlFor="hoverOff" className="hoverOff">{lbltext}</label>
+                <input id="hoverOff" type="checkbox" onClick={changeHoverStatus} onChange={() => setHoverStatus(!hoverStatus)}></input><br></br>
+                <label className="bscardslbl" htmlFor="bscards">Show messages in Bootstrap cards</label>
+                <input id="bscards" type="checkbox" onChange={changeViews}></input>
+            </span>
 
-            {showInCards && <MessageCards messages={messages} msgIcon={msgIcon}/>}
+
+
+            {showInCards && <MessageCards messages={messages} msgIcon={msgIcon} />}
 
 
             <div className="msg" hidden={hideMsgDiv}>
@@ -204,37 +250,46 @@ const Messages = (props) => {
                 lailla kuin esim i for loopissa*/}
                 {messages.map(message => (
                     //divin classnamea voidaan muuttaa state-muuttujan avulla.
-                    <div className={hoverOff} key={message.id} onMouseEnter={()=>starsSelected(message.id)} >
-                        {/*ikoneita pystyy käyttämään silmukassa*/}
-
+                    <div className={hoverOff} key={message.id} onMouseEnter={() => starsSelected(message.id)} >
+                       {/*talletetaan msgtxt:n ensimmäinen kirjain muuttujaan*/}
+                        <p hidden>{first=message.msgtxt[0]}</p>
+                         {/*ikoneita pystyy käyttämään silmukassa*/}
                         <img src={msgIcon} alt="icon"></img>
                         <p hidden={hideidAndDate}>Message id: {message.id}</p>
-                        <p id={"c"+message.id}></p>
-                        <p id={"m"+message.id} className={"m"+message.id}>{message.msgtxt}</p>
-                        <p hidden={hideidAndDate} className="msgTime">Posting time: <b>{message.txtposttime}</b></p>
+                        <p id={"c" + message.id}></p>
+                        {/*jokaisen viestin ensimmäinen kirjain first-muuttujassa ja kirjain näytetään 200% suurempana kuin muut replacella korvataan
+                        ensimmäinen kirjain tyhjällä koska muuten eka kirjain näytettäisiin kahdesti*/}
+                        <p id={"m" + message.id} className={"m" + message.id}><p className="firstletter">{first}</p>{message.msgtxt.replace(first,"")}</p>
+                        <p hidden={hideidAndDate} className="msgTime">Posting time: <b>{message.txtposttime.replace(".", "/").replace(".", "/")}</b></p>
                         <p hidden={hideidAndDate}>Likes: {message.likes}</p>
                         <p hidden={hideidAndDate}>Unlikes: {message.unlike}</p>
-                       {/*yksilöidään span elementit message.id:n avulla että voidaan päivittää
+                        <p>Characters: {message.msgtxt.length}</p>
+                        
+
+
+
+
+                        {/*yksilöidään span elementit message.id:n avulla että voidaan päivittää
                        aina vain halutun viestin ratingia*/}
-                        <span id={"star1"+message.id} class="fa fa-star" onClick={()=>rate(message.id)} ></span>
-                        <span id={"star2"+message.id} class="fa fa-star" onClick={()=>rate(message.id)} ></span>
-                        <span id={"star3"+message.id} class="fa fa-star" onClick={()=>rate(message.id)} ></span>
-                        <span id={"star4"+message.id} class="fa fa-star" onClick={()=>rate(message.id)} ></span>
-                        <span id={"star5"+message.id} class="fa fa-star" onClick={()=>rate(message.id)} ></span>
-                        
-                        
-                        
+                        <span id={"star1" + message.id} class="fa fa-star" onClick={() => rate(message.id)} ></span>
+                        <span id={"star2" + message.id} class="fa fa-star" onClick={() => rate(message.id)} ></span>
+                        <span id={"star3" + message.id} class="fa fa-star" onClick={() => rate(message.id)} ></span>
+                        <span id={"star4" + message.id} class="fa fa-star" onClick={() => rate(message.id)} ></span>
+                        <span id={"star5" + message.id} class="fa fa-star" onClick={() => rate(message.id)} ></span>
+
+
+
                         <div className="crudBtns">
                             {/*lähetetään postauksen id numero handleDelete fuktiolle*/}
-                            <button disabled={props.delBtnDisable} onClick={() => handleDelete(message.id)}>
-                            <img className="alarm" src={alert} ></img>
-                               <img src={binIcon} alt="remove icon"></img>
+                            <button disabled={props.delBtnDisable} onClick={() => handleDelete(message.id, message.msgtxt)}>
+                                <img className="alarm" src={alertIcon} ></img>
+                                <img src={binIcon} alt="remove icon"></img>
                             </button>
                             <button onClick={() => handleLike(message.id)}>
                                 {/*png kuvan sisällytys button elementtiin.*/}
                                 <img src={likeIcon} alt="like icon"></img>
                             </button>
-                           
+
                             <button className="delbtn" onClick={() => handleUnLike(message.id)}>
                                 <img src={unlikeIcon}></img>
                             </button>
@@ -243,17 +298,17 @@ const Messages = (props) => {
                             sisällä, että linkitys toimii klikattaessa*/}
                                 <Link to={`/update/${message.id}`}><img src={updateIcon}></img></Link>
                             </button>
-                            <button className="censor" onClick={()=>DoCensor(message.id)}>
+                            <button className="censor" onClick={() => DoCensor(message.id)}>
                                 <img src={censor}></img>
                             </button>
-                            <button onClick={()=>resetStars(message.id)}><img src={reset}></img></button>
+                            <button onClick={() => resetStars(message.id)}><img src={reset}></img></button>
                         </div>
 
                     </div>
 
 
                 ))}
-             
+
 
 
                 {/*viestien kokonaismäärä näytetään map silmukan ulkopuolella, muuten se tulostuisi
